requirements.txt
pandas
joblib
streamlit
scikit-learn
langchain
langchain-groq
python-dotenv